symbolic/requirements.txt
symbolica-agentica>=0.1.0
python-dotenv>=1.0.0

